+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog 
 
 ## 3.0.4
+* fixed float for dataTables
 
 ## 3.0.3
 * New widget [Presenter](http://metroui.org.ua/presenter.html)